adapt req hooks
