+++ packages/insomnia/src/common/har.ts
@@ -1,14 +1,10 @@
-import clone from 'clone';
 import type * as Har from 'har-format';
 import { Cookie as ToughCookie } from 'tough-cookie';
 
 import type { BaseModel, Environment, Request, RequestGroup, Response, Workspace } from '~/insomnia-data';
 import { models, services } from '~/insomnia-data';
+import { applyRequestHooks } from '~/network/network-adapter';
 
-import * as plugins from '../plugins';
-import * as pluginApp from '../plugins/context/app';
-import * as pluginRequest from '../plugins/context/request';
-import * as pluginStore from '../plugins/context/store';
 import { RenderError } from '../templating/render-error';
 import type { RenderedRequest } from '../templating/types';
 import { parseGraphQLReqeustBody } from '../utils/graph-ql';
@@ -248,7 +244,7 @@ export async function exportHarRequest(requestId: string, environmentOrWorkspace
 export async function exportHarWithRequest(request: Request, environmentId?: string, addContentLength = false) {
   try {
     const renderResult = await getRenderedRequestAndContext({ request, environment: environmentId });
-    const renderedRequest = await _applyRequestPluginHooks(renderResult.request, renderResult.context);
+    const renderedRequest = await applyRequestHooks(renderResult.request, renderResult.context);
     parseGraphQLReqeustBody(renderedRequest);
     return exportHarWithRenderedRequest(renderedRequest, addContentLength);
   } catch (err) {
@@ -258,31 +254,6 @@ export async function exportHarWithRequest(request: Request, environmentId?: str
         : new Error(`Failed to export request "${request.name}"\n ${err.message}`);
     throw error;
   }
-}
-
-async function _applyRequestPluginHooks(
-  renderedRequest: RenderedRequest,
-  renderedContext: Record<string, any>,
-): Promise<RenderedRequest> {
-  let newRenderedRequest = renderedRequest;
-
-  for (const { plugin, hook } of await plugins.getRequestHooks()) {
-    newRenderedRequest = clone(newRenderedRequest);
-    const context = {
-      ...(pluginApp.init() as Record<string, any>),
-      ...(pluginRequest.init(newRenderedRequest, renderedContext) as Record<string, any>),
-      ...(pluginStore.init(plugin) as Record<string, any>),
-    };
-
-    try {
-      await hook(context);
-    } catch (err) {
-      err.plugin = plugin;
-      throw err;
-    }
-  }
-
-  return newRenderedRequest;
 }
 
 export async function exportHarWithRenderedRequest(renderedRequest: RenderedRequest, addContentLength = false) {
