+++ packages/insomnia/src/network/network-adapter.renderer.ts
@@ -1,3 +1,6 @@
+import clone from 'clone';
+
+import { filterHeaders } from '~/common/misc';
 import type { RequestHeader } from '~/insomnia-data';
 import { plugins as pluginsBridge } from '~/plugins/renderer-bridge';
 import type { RenderedRequest } from '~/templating/types';
@@ -7,8 +10,7 @@ import type { CurlRequestOptions, ResponsePatch } from '../main/network/libcurl-
 import { cancellableCurlRequest } from './cancellation';
 import { runScriptConcurrently } from './concurrency';
 
-export const getTimelinePath = (responseId: string): Promise<string> =>
-  window.main.timeline.getPath(responseId);
+export const getTimelinePath = (responseId: string): Promise<string> => window.main.timeline.getPath(responseId);
 
 export const appendToTimelineOnError = (timelinePath: string, data: string): Promise<void> =>
   window.main.timeline.appendToFile({ timelinePath, data });
@@ -19,8 +21,7 @@ export const appendTimelineLines = (timelinePath: string, logs: string[]): Promi
 export const getAuthHeader = (r: RenderedRequest, u: string): Promise<RequestHeader | undefined> =>
   window.main.getAuthHeader(r, u);
 
-export const executeCurlRequest = (options: CurlRequestOptions) =>
-  cancellableCurlRequest(options);
+export const executeCurlRequest = (options: CurlRequestOptions) => cancellableCurlRequest(options);
 
 export const runScript = (options: {
   script: string;
@@ -31,11 +32,28 @@ export async function applyRequestHooks(
   newRenderedRequest: RenderedRequest,
   renderedContext: Record<string, any>,
 ): Promise<RenderedRequest> {
-  if (!await pluginsBridge.hasRequestHooks()) {
-    return newRenderedRequest;
+  const request = clone(newRenderedRequest);
+  const defaultHeaders = renderedContext['DEFAULT_HEADERS'];
+  if (defaultHeaders && typeof defaultHeaders === 'object' && !Array.isArray(defaultHeaders)) {
+    for (const name of Object.keys(defaultHeaders)) {
+      const value = (defaultHeaders as Record<string, any>)[name];
+      const existing = filterHeaders(request.headers, name);
+      if (existing.length) {
+        console.log(`[header] Skip setting default header ${name}. Already set to ${value}`);
+      } else if (value === 'null') {
+        request.headers = request.headers.filter(h => filterHeaders([h], name).length === 0);
+        console.log(`[header] Remove default header ${name}`);
+      } else {
+        request.headers.push({ name, value: String(value) });
+        console.log(`[header] Set default header ${name}: ${value}`);
+      }
+    }
+  }
+  if (!(await pluginsBridge.hasRequestHooks())) {
+    return request;
   }
   return pluginsBridge.applyRequestHooks({
-    renderedRequest: newRenderedRequest,
+    renderedRequest: request,
     projectId: renderedContext.getProjectId(),
     environment: renderedContext,
   });
@@ -46,7 +64,7 @@ export async function applyResponseHooks(
   renderedRequest: RenderedRequest,
   renderedContext: Record<string, any>,
 ): Promise<ResponsePatch> {
-  if (!await pluginsBridge.hasResponseHooks()) {
+  if (!(await pluginsBridge.hasResponseHooks())) {
     return response;
   }
   return pluginsBridge.applyResponseHooks({
