+++ packages/insomnia/src/network/network.ts
@@ -1,4 +1,3 @@
-import clone from 'clone';
 import orderedJSON from 'json-order';
 
 import type {
@@ -45,7 +44,6 @@ import { generateId, getContentTypeHeader, getLocationHeader, getSetCookieHeader
 import { getRenderedRequestAndContext } from '../common/render';
 import { ascendingFirstIndexStringSort } from '../common/sorting';
 import type { HeaderResult, ResponsePatch } from '../main/network/libcurl-promise';
-import * as pluginRequest from '../plugins/context/request';
 import { RenderError } from '../templating/render-error';
 import type { RenderedRequest, RenderPurpose } from '../templating/types';
 import { maskOrDecryptVaultDataIfNecessary } from '../templating/utils';
@@ -799,7 +797,7 @@ export const tryToTransformRequestWithPlugins = async (renderResult: {
 }) => {
   const { request, context } = renderResult;
   try {
-    return await _applyRequestPluginHooks(request, context);
+    return await applyRequestHooks(request, context);
   } catch {
     throw new Error(`Failed to transform request with plugins: ${request._id}`);
   }
@@ -951,7 +949,19 @@ export const responseTransform = async (
     return response;
   }
   console.log(`[network] Response succeeded req=${patch.parentId} status=${response.statusCode || '?'}`);
-  return await _applyResponsePluginHooks(response, renderedRequest, context);
+  try {
+    return await applyResponseHooks(response, renderedRequest, context);
+  } catch (err) {
+    console.log('[plugin] Response hook failed', err, response);
+    return {
+      url: renderedRequest.url,
+      error: `[plugin] Response hook failed err=${err instanceof Error ? err.message : String(err)}`,
+      elapsedTime: 0,
+      statusMessage: 'Error',
+      settingSendCookies: renderedRequest.settingSendCookies,
+      settingStoreCookies: renderedRequest.settingStoreCookies,
+    };
+  }
 };
 export function getAuthQueryParams(authentication: RequestAuthentication) {
   if (authentication.disabled) {
@@ -1040,50 +1050,6 @@ export const getCurrentUrl = ({ headerResults, finalUrl }: { headerResults: any;
     return finalUrl;
   }
 };
-
-export async function _applyRequestPluginHooks(renderedRequest: RenderedRequest, renderedContext: Record<string, any>) {
-  const newRenderedRequest = clone(renderedRequest);
-
-  // Apply built-in default-headers hook in the renderer (no IPC needed)
-  const { request: reqCtx } = pluginRequest.init(newRenderedRequest, renderedContext);
-  const defaultHeaders = reqCtx.getEnvironmentVariable('DEFAULT_HEADERS');
-  if (defaultHeaders && typeof defaultHeaders === 'object' && !Array.isArray(defaultHeaders)) {
-    for (const name of Object.keys(defaultHeaders)) {
-      const value = (defaultHeaders as Record<string, any>)[name];
-      if (reqCtx.hasHeader(name)) {
-        console.log(`[header] Skip setting default header ${name}. Already set to ${value}`);
-      } else if (value === 'null') {
-        reqCtx.removeHeader(name);
-        console.log(`[header] Remove default header ${name}`);
-      } else {
-        reqCtx.setHeader(name, value);
-        console.log(`[header] Set default header ${name}: ${value}`);
-      }
-    }
-  }
-
-  return applyRequestHooks(newRenderedRequest, renderedContext);
-}
-
-export async function _applyResponsePluginHooks(
-  response: ResponsePatch,
-  renderedRequest: RenderedRequest,
-  renderedContext: Record<string, any>,
-): Promise<ResponsePatch> {
-  try {
-    return await applyResponseHooks(response, renderedRequest, renderedContext);
-  } catch (err) {
-    console.log('[plugin] Response hook failed', err, response);
-    return {
-      url: renderedRequest.url,
-      error: `[plugin] Response hook failed err=${err instanceof Error ? err.message : String(err)}`,
-      elapsedTime: 0, // 0 because this path is hit during plugin calls
-      statusMessage: 'Error',
-      settingSendCookies: renderedRequest.settingSendCookies,
-      settingStoreCookies: renderedRequest.settingStoreCookies,
-    };
-  }
-}
 
 export const defaultSendActionRuntime: SendActionRuntime = {
   appendTimeline: appendTimelineLines,
